Chore: move background tracker for thread responses
